cvar.c: 	%g is no good :/ menu.c screen.c: 	float -> int cleanup for scr_viewsize vid_glx.c: 	remove unused vid_mode cvar

+++ source/screen.c
@@ -979,7 +979,7 @@ needs almost the entire 256k of stack space!
 */
 void SCR_UpdateScreen (void)
 {
-	static float	oldscr_viewsize;
+	static int	oldscr_viewsize;
 	vrect_t		vrect;
 
 	if (scr_skipupdate || block_drawing)
